LayoutTests/platform/ios-sim/svg/custom/getSubStringLength-expected.txt

PASS svgText.getSubStringLength(-1, 2) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgText.getSubStringLength(-1, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
FAIL svgText.getSubStringLength(1, 3) should be 40. Was 40.000003814697266.
FAIL svgText.getSubStringLength(0, 4) should be 60. Was 60.00000762939453.
PASS svgText.getSubStringLength(3, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgText.getSubStringLength(0, 0) is 0
PASS svgText.getSubStringLength(2, 0) is 0
FAIL svgText.getSubStringLength(0, 1) should be 20. Was 20.000001907348633.
FAIL svgText.getSubStringLength(1, 1) should be 20. Was 20.000001907348633.
FAIL svgText.getSubStringLength(2, 1) should be 20. Was 20.000001907348633.
FAIL svgText.getSubStringLength(0, 3) should be 60. Was 60.00000762939453.
PASS svgText.getSubStringLength(1, -1) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgText.getSubStringLength(2, -1) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgText.getSubStringLength(3, -1) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgText.getSubStringLength(3, -3) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS successfullyParsed is true

TEST COMPLETE
